eatalyser-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
requests
pillow
python-multipart
tensorflow-datasets
matplotlib
tensorflow
tensorflow-hub
tensorflow-macos
tensorflow-metal
opencv-python
passlib[bcrypt]
python-jose
pydantic[email]
torch
torchvision
scikit-learn
python-jose
python-multipart



